All non-function styling moved to <colgroup>
