+++ addon/styles/addon.scss
@@ -1,11 +1,14 @@
 .data-grid {
   table {
+    table-layout:fixed;
     border-collapse: collapse;
     width: 100%;
 
     th, td {
       padding-left: 5px;
       padding-right: 1em;
+      text-overflow: ellipsis;
+      overflow: hidden;
     }
 
     thead {
