+++ addon/templates/components/data-row.hbs
@@ -1,10 +1,10 @@
 {{#if hasActions}}
-  <td style="width:15px;">
+  <td>
     {{check-box value=isSelected
                 object=item
                 updated=selectItem}}
   </td>
 {{/if}}
 {{#each properties as |property|}}
-  {{data-col property=property item=item}}
+  {{data-cell property=property item=item}}
 {{/each}}
