+++ app/components/data-cell.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-railio-grid/components/data-cell';
