--- app/components/data-col.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-railio-grid/components/data-col';
