+++ app/components/data-column.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-railio-grid/components/data-column';
