+++ app/templates/components/data-column.hbs
@@ -1,0 +1,1 @@
+{{value}}
